requirements.txt
--index-url https://mirrors.aliyun.com/pypi/simple

aiofiles==23.1.0 ; python_version >= "3.9" and python_version < "4.0"
aiohttp==3.8.5 ; python_version >= "3.9" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
aiosqlite==0.17.0 ; python_version >= "3.9" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.9" and python_version < "4.0"
apscheduler==3.10.1 ; python_version >= "3.9" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.9" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.9" and python_version < "4.0"
cachetools==5.3.1 ; python_version >= "3.9" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.2.0 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.6 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
deprecated==1.2.14 ; python_version >= "3.9" and python_version < "4.0"
emoji==2.7.0 ; python_version >= "3.9" and python_version < "4.0"
exceptiongroup==1.1.2 ; python_version >= "3.9" and python_version < "3.11"
fastapi==0.100.1 ; python_version >= "3.9" and python_version < "4.0"
frozenlist==1.4.0 ; python_version >= "3.9" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
httpcore==0.17.3 ; python_version >= "3.9" and python_version < "4.0"
httptools==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
httpx==0.24.1 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
iso8601==1.1.0 ; python_version >= "3.9" and python_version < "4.0"
loguru==0.7.0 ; python_version >= "3.9" and python_version < "4.0"
msgpack==1.0.5 ; python_version >= "3.9" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.9" and python_version < "4.0"
nonebot-adapter-kaiheila==0.2.8 ; python_version >= "3.9" and python_version < "4.0"
nonebot-adapter-onebot==2.2.3 ; python_version >= "3.9" and python_version < "4.0"
nonebot-adapter-qqguild==0.2.3 ; python_version >= "3.9" and python_version < "4.0"
nonebot-plugin-apscheduler==0.3.0 ; python_version >= "3.9" and python_version < "4.0"
nonebot-plugin-send-anything-anywhere==0.2.7 ; python_version >= "3.9" and python_version < "4.0"
nonebot-plugin-session==0.0.9 ; python_version >= "3.9" and python_version < "4.0"
nonebot-plugin-userinfo==0.0.2 ; python_version >= "3.9" and python_version < "4.0"
nonebot2==2.0.1 ; python_version >= "3.9" and python_version < "4.0"
nonebot2[fastapi]==2.0.1 ; python_version >= "3.9" and python_version < "4.0"
pillow==9.5.0 ; python_version >= "3.9" and python_version < "4.0"
pydantic==1.10.12 ; python_version >= "3.9" and python_version < "4.0"
pydantic[dotenv]==1.10.12 ; python_version >= "3.9" and python_version < "4.0"
pygtrie==2.5.0 ; python_version >= "3.9" and python_version < "4.0"
pypika-tortoise==0.1.6 ; python_version >= "3.9" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.9" and python_version < "4.0"
setuptools==68.0.0 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.9" and python_version < "4.0"
strenum==0.4.15 ; python_version >= "3.9" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.9" and python_version < "3.11"
tortoise-orm==0.19.3 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.9" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
tzlocal==5.0.1 ; python_version >= "3.9" and python_version < "4.0"
uvicorn[standard]==0.23.2 ; python_version >= "3.9" and python_version < "4.0"
uvloop==0.17.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.9" and python_version < "4.0"
watchfiles==0.19.0 ; python_version >= "3.9" and python_version < "4.0"
websockets==11.0.3 ; python_version >= "3.9" and python_version < "4.0"
win32-setctime==1.1.0 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
wrapt==1.15.0 ; python_version >= "3.9" and python_version < "4.0"
yarl==1.9.2 ; python_version >= "3.9" and python_version < "4.0"
